post-proc scripts adjusted to account for multiple input files and comparison
